JDF-335 Modified project structure to Hybrid Mobile project.
This quickstart now uses the Hybrid Mobile project structure in
JBDS 7. The project can be imported into JBDS 7 as a Hybrid
Mobile project and deployed to Android or iOS  devices or
emulators.

Note - this forcibly adds an Eclipse project so that it can be
shared among users and imported into JBDS 7.1 workspaces.

Synchronized the code in this quickstart with the base quickstart.

The code is more or less in sync with kitchensink-html5-mobile.