JDF-579 Added a quickstart for Cordova Contacts API.
This quickstart extends the kitchensink-backbone quickstart,
adding in capabilities to demonstrate the use of the Cordova
Contacts API. The Contacts API is used to demonstrate searching
of contacts present in the device, as well as saving of contacts
onto the device.

This quickstart is now synchronized with the base quickstart.
This is now up to date with the kitchensink-backbone quickstart.